code/python_app/requirements.txt
# OTel Mandatory Packages 
opentelemetry-sdk
opentelemetry-instrumentation

# OTel Instrumentation Packages for flask and python
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests

# OTel Exporter Packages 
opentelemetry-exporter-otlp

# Python App 
flask 
requests

# Productionization
gunicorn